Move CharSequence utility functions to 'base'
